backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic-settings==2.10.1
celery==5.5.3
redis==6.4.0
httpx==0.28.1
msal==1.31.1
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.4.1
